fix(Input): should not have value prop on input when click toggle icon
